+++ Athena/Source/Athena/Math/TypesImpl/Quaternion.h
@@ -41,7 +41,7 @@ namespace Athena::Math
 
 // -------------Public Methods-------------------------------------
 	public:
-		constexpr SIZE_T Size() const
+		constexpr uint32 Size() const
 		{
 			return 4;
 		}
@@ -130,13 +130,13 @@ namespace Athena::Math
 
 // -------------Operators-------------------------------------
 	public:
-		constexpr const T& operator[](SIZE_T idx) const
+		constexpr const T& operator[](uint32 idx) const
 		{
 			ATN_CORE_ASSERT(idx < 4, "Quaternion subscript out of range");
 			return *(&w + idx);
 		}
 
-		constexpr T& operator[](SIZE_T idx)
+		constexpr T& operator[](uint32 idx)
 		{
 			ATN_CORE_ASSERT(idx < 4, "Quaternion subscript out of range");
 			return *(&w + idx);
